requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nanoid==2.0.0
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pylint==3.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
starlette-session==0.4.3
tomlkit==0.12.4
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
